+++ src/cli/crawl-specs.js
@@ -29,6 +29,7 @@ const requireFromWorkingDirectory = require('../lib/util').requireFromWorkingDir
 const completeWithInfoFromW3CApi = require('../lib/util').completeWithInfoFromW3CApi;
 const completeWithShortName = require('../lib/util').completeWithShortName;
 const getShortname = require('../lib/util').getShortname;
+const isLatestLevel = require('../lib/util').isLatestLevel;
 const processSpecification = require('../lib/util').processSpecification;
 
 /**
@@ -106,13 +107,15 @@ function completeWithInfoFromSpecref(specs) {
 async function createInitialSpecDescriptions(list) {
     function createSpecObject(spec) {
         let res = {
-            url: (typeof spec === 'string') ? spec : (spec.url || 'about:blank')
+            url: (typeof spec === 'string') ? spec.split(' ')[0] : (spec.url || 'about:blank')
         };
         if ((typeof spec !== 'string') && spec.html) {
             res.html = spec.html;
         }
         res.flags = {
-            delta: !!spec.delta
+            delta: (typeof spec === 'string') ?
+                spec.split(' ').includes('delta') :
+                !!spec.delta
         };
         return res;
     }
@@ -367,38 +370,11 @@ async function saveResults(crawlInfo, crawlOptions, data, folder) {
             }));
     };
 
-    // Helper function that returns true when the given spec is is the latest
-    // level of that spec in the crawl for the given type of content
-    // ("css" or "idl"). Note the code handles the special case of the CSS2
-    // and CSS22 specs, and assumes that URLs that don't end with a level
-    // number are at level 1 (this does not work for CSS specs whose URLs still
-    // follow the old `css3-` pattern, but we're only interested in comparing
-    // with more recent levels in that case, so it does not matter)
-    const isLatestLevel = spec => {
-        const getLevel = spec =>
-            (spec.url.match(/-\d+\/$/) ?
-            parseInt(spec.url.match(/-(\d+)\/$/)[1], 10) :
-            (spec.url.match(/CSS22\/$/i) ? 2 : 1));
-        const shortname = getShortname(spec);
-        const level = getLevel(spec);
-        const candidates = data.filter(s => !s.flags.delta &&
-            (getShortname(s) === shortname) &&
-            (getLevel(s) >= level));
-
-        // Note the list of candidates for this shortname includes the spec
-        // itself. It is the latest level if there is no other candidate at
-        // a strictly greater level, and if the spec under consideration is
-        // the first element in the list (for the hopefully rare case where
-        // we have two candidate specs that are at the same level)
-        return !candidates.find(s => getLevel(s) > level) &&
-            (candidates[0] === spec);
-    };
-
     // Save IDL dumps for the latest level of a spec to the idl folder
     await Promise.all(data
         .filter(spec => !spec.flags.delta)
         .filter(spec => spec.idl && spec.idl.idl)
-        .filter(spec => isLatestLevel(spec))
+        .filter(spec => isLatestLevel(spec, data))
         .map(saveIdl));
 
     // Save CSS dumps for the latest level of a spec to the css folder
@@ -408,7 +384,7 @@ async function saveResults(crawlInfo, crawlOptions, data, folder) {
             (Object.keys(spec.css.properties || {}).length > 0) ||
             (Object.keys(spec.css.descriptors || {}).length > 0) ||
             (Object.keys(spec.css.valuespaces || {}).length > 0)))
-        .filter(spec => isLatestLevel(spec))
+        .filter(spec => isLatestLevel(spec, data))
         .map(saveCss));
 
     // Save all results to the crawl.json file
@@ -453,9 +429,6 @@ function assembleListOfSpec(filename, nested) {
         .map(u => (typeof u === 'string') ? Object.assign({ url: u }) : u)
         .map(u => u.file ? assembleListOfSpec(path.resolve(path.dirname(filename), u.file), true) : u);
     crawlInfo.list = flatten(crawlInfo.list);
-    if (filename.match(/-delta/)) {
-        crawlInfo.list.forEach(u => u.delta = true);
-    }
     crawlInfo.list = crawlInfo.list.filter(u => {
         const first = crawlInfo.list.find(s => s.url === u.url);
         return first === u;
