+++ src/specs/specs-css.json
@@ -1,63 +1,28 @@
 [
   "https://compat.spec.whatwg.org/",
   "https://drafts.css-houdini.org/box-tree-api/",
+  "https://drafts.css-houdini.org/css-typed-om-2/ delta",
   "https://drafts.css-houdini.org/font-metrics-api/",
-  {
-    "url": "https://drafts.csswg.org/css-animations-2/",
-    "delta": true
-  },
-  {
-    "url": "https://drafts.csswg.org/css-backgrounds-4/",
-    "delta": true
-  },
+  "https://drafts.csswg.org/css-animations-2/ delta",
+  "https://drafts.csswg.org/css-backgrounds-4/ delta",
   "https://drafts.csswg.org/css-box-3/",
-  {
-    "url": "https://drafts.csswg.org/css-conditional-4/",
-    "delta": true
-  },
+  "https://drafts.csswg.org/css-conditional-4/ delta",
   "https://drafts.csswg.org/css-env-1/",
   "https://drafts.csswg.org/css-exclusions-1/",
-  {
-    "url": "https://drafts.csswg.org/css-gcpm-4/",
-    "delta": true
-  },
-  {
-    "url": "https://drafts.csswg.org/css-multicol-2/",
-    "delta": true
-  },
+  "https://drafts.csswg.org/css-gcpm-4/ delta",
+  "https://drafts.csswg.org/css-multicol-2/ delta",
   "https://drafts.csswg.org/css-nesting-1/",
-  {
-    "url": "https://drafts.csswg.org/css-page-4/",
-    "delta": true
-  },
-  {
-    "url": "https://drafts.csswg.org/css-shapes-2/",
-    "delta": true
-  },
+  "https://drafts.csswg.org/css-page-4/ delta",
+  "https://drafts.csswg.org/css-shapes-2/ delta",
   "https://drafts.csswg.org/css-size-adjust-1/",
-  {
-    "url": "https://drafts.csswg.org/css-sizing-4/",
-    "delta": true
-  },
-  {
-    "url": "https://drafts.csswg.org/css-transforms-2/",
-    "delta": true
-  },
-  {
-    "url": "https://drafts.csswg.org/css-transitions-2/",
-    "delta": true
-  },
+  "https://drafts.csswg.org/css-sizing-4/ delta",
+  "https://drafts.csswg.org/css-transforms-2/ delta",
+  "https://drafts.csswg.org/css-transitions-2/ delta",
   "https://drafts.csswg.org/css-tv/",
-  {
-    "url": "https://drafts.csswg.org/mediaqueries-5/",
-    "delta": true
-  },
+  "https://drafts.csswg.org/mediaqueries-5/ delta",
   "https://drafts.csswg.org/scroll-animations-1/",
   "https://drafts.fxtf.org/compositing-2/",
-  {
-    "url": "https://drafts.fxtf.org/filter-effects-2/",
-    "delta": true
-  },
+  "https://drafts.fxtf.org/filter-effects-2/ delta",
   "https://quirks.spec.whatwg.org/",
   "https://www.w3.org/TR/compositing-1/",
   "https://www.w3.org/TR/cssom-view-1/",
@@ -66,19 +31,13 @@
   "https://www.w3.org/TR/css-animation-worklet-1/",
   "https://www.w3.org/TR/css-backgrounds-3/",
   "https://www.w3.org/TR/css-break-3/",
-  {
-    "url": "https://www.w3.org/TR/css-break-4/",
-    "delta": true
-  },
+  "https://www.w3.org/TR/css-break-4/ delta",
   "https://www.w3.org/TR/css-cascade-3/",
   "https://www.w3.org/TR/css-cascade-4/",
   "https://www.w3.org/TR/css-color-adjust-1/",
   "https://www.w3.org/TR/css-color-3/",
   "https://www.w3.org/TR/css-color-4/",
-  {
-    "url": "https://www.w3.org/TR/css-color-5/",
-    "delta": true
-  },
+  "https://www.w3.org/TR/css-color-5/ delta",
   "https://www.w3.org/TR/css-contain-1/",
   "https://www.w3.org/TR/css-content-3/",
   "https://www.w3.org/TR/css-counter-styles-3/",
@@ -91,14 +50,8 @@
   "https://www.w3.org/TR/css-fonts-4/",
   "https://www.w3.org/TR/css-gcpm-3/",
   "https://www.w3.org/TR/css-grid-1/",
-  {
-    "url": "https://www.w3.org/TR/css-grid-2/",
-    "delta": true
-  },
-  {
-    "url": "https://www.w3.org/TR/css-images-4/",
-    "delta": true
-  },
+  "https://www.w3.org/TR/css-grid-2/ delta",
+  "https://www.w3.org/TR/css-images-4/ delta",
   "https://www.w3.org/TR/css-inline-3/",
   "https://www.w3.org/TR/css-layout-api-1/",
   "https://www.w3.org/TR/css-line-grid-1/",
@@ -109,10 +62,7 @@
   "https://www.w3.org/TR/css-namespaces-3/",
   "https://www.w3.org/TR/css-nav-1/",
   "https://www.w3.org/TR/css-overflow-3/",
-  {
-    "url": "https://www.w3.org/TR/css-overflow-4/",
-    "delta": true
-  },
+  "https://www.w3.org/TR/css-overflow-4/ delta",
   "https://www.w3.org/TR/css-overscroll-1/",
   "https://www.w3.org/TR/css-page-3/",
   "https://www.w3.org/TR/css-page-floats-3/",
@@ -135,15 +85,9 @@
   "https://www.w3.org/TR/css-syntax-3/",
   "https://www.w3.org/TR/css-tables-3/",
   "https://www.w3.org/TR/css-text-3/",
-  {
-    "url": "https://www.w3.org/TR/css-text-4/",
-    "delta": true
-  },
+  "https://www.w3.org/TR/css-text-4/ delta",
   "https://www.w3.org/TR/css-text-decor-3/",
-  {
-    "url": "https://www.w3.org/TR/css-text-decor-4/",
-    "delta": true
-  },
+  "https://www.w3.org/TR/css-text-decor-4/ delta",
   "https://www.w3.org/TR/css-transforms-1/",
   "https://www.w3.org/TR/css-transitions-1/",
   "https://www.w3.org/TR/css-typed-om-1/",
